src/requirements.txt
fastapi 
uvicorn
pillow
numpy
python-multipart
prometheus-fastapi-instrumentator
prometheus_client
psutil